feat: renaming and minor changes
